+++ templates/workspaces/base/terraform/backup/variables.tf
@@ -7,6 +7,9 @@ variable "tre_id" {
 variable "resource_group_name" {
   type = string
 }
+variable "resource_group_id" {
+  type = string
+}
 variable "tre_workspace_tags" {
   type = map(string)
 }
@@ -15,4 +18,12 @@ variable "tre_resource_id" {
 }
 variable "enable_cmk_encryption" {
   type = bool
+}
+variable "encryption_key_versionless_id" {
+  type    = string
+  default = null
+}
+variable "encryption_identity_id" {
+  type    = string
+  default = null
 }