+++ templates/workspaces/base/terraform/workspace.tf
@@ -107,12 +107,15 @@ module "azure_monitor" {
 }
 
 module "backup" {
-  count                 = var.enable_backup ? 1 : 0
-  source                = "./backup"
-  tre_id                = var.tre_id
-  tre_resource_id       = var.tre_resource_id
-  location              = var.location
-  resource_group_name   = azurerm_resource_group.ws.name
-  tre_workspace_tags    = local.tre_workspace_tags
-  enable_cmk_encryption = var.enable_cmk_encryption
+  count                         = var.enable_backup ? 1 : 0
+  source                        = "./backup"
+  tre_id                        = var.tre_id
+  tre_resource_id               = var.tre_resource_id
+  location                      = var.location
+  resource_group_name           = azurerm_resource_group.ws.name
+  resource_group_id             = azurerm_resource_group.ws.id
+  tre_workspace_tags            = local.tre_workspace_tags
+  enable_cmk_encryption         = var.enable_cmk_encryption
+  encryption_key_versionless_id = var.enable_cmk_encryption ? azurerm_key_vault_key.encryption_key[0].versionless_id : null
+  encryption_identity_id        = var.enable_cmk_encryption ? azurerm_user_assigned_identity.encryption_identity[0].id : null
 }